refactor(CC-742): extract city-switch navigation into useCitySwitchNavigation hook
Same resolveCitySwitchPath + toast + router.push logic was duplicated in
navigation-bar.tsx and twice in JNDrawer.tsx; addresses PR review comment.
